requirements.txt
aws-cdk-lib==2.139.0
constructs>=10.0.0,<11.0.0
cdk-nag>=2.28.67
PyYAML>=6.0
